# improvement: avoid ub and delele unnecessary unsafe