--- a/common/exception/src/exception_code.rs
+++ b/common/exception/src/exception_code.rs
@@ -144,6 +144,9 @@ build_exceptions! {
     // Tenant error codes.
     TenantIsEmpty(1101),
     IndexOutOfBounds(1102),
+
+    // Layout error code.
+    LayoutError(1103),
 }
 
 // Metasvr errors [2001, 3000].