--- a/query/src/pipelines/new/processors/transforms/aggregator/aggregator_params.rs
+++ b/query/src/pipelines/new/processors/transforms/aggregator/aggregator_params.rs
@@ -40,7 +40,8 @@ pub struct AggregatorParams {
     pub aggregate_functions_arguments_name: Vec<Vec<String>>,
 
     // about function state memory layout
-    pub layout: Layout,
+    // If there is no aggregate function, layout is None
+    pub layout: Option<Layout>,
     pub offsets_aggregate_states: Vec<usize>,
 }
 
@@ -65,7 +66,15 @@ impl AggregatorParams {
             aggregate_functions_arguments_name.push(expr.to_aggregate_function_names()?);
         }
 
-        let (states_layout, states_offsets) = unsafe { get_layout_offsets(&aggregate_functions) };
+        let mut states_offsets: Vec<usize> = Vec::with_capacity(aggregate_functions.len());
+        let mut states_layout = None;
+        if !aggregate_functions.is_empty() {
+            states_offsets = Vec::with_capacity(aggregate_functions.len());
+            states_layout = Option::Some(get_layout_offsets(
+                aggregate_functions.as_slice(),
+                &mut states_offsets,
+            )?);
+        }
 
         let group_data_fields = plan
             .group_expr
@@ -103,7 +112,15 @@ impl AggregatorParams {
             aggregate_functions_arguments_name.push(expr.to_aggregate_function_names()?);
         }
 
-        let (states_layout, states_offsets) = unsafe { get_layout_offsets(&aggregate_functions) };
+        let mut states_offsets: Vec<usize> = Vec::with_capacity(aggregate_functions.len());
+        let mut states_layout = None;
+        if !aggregate_functions.is_empty() {
+            states_offsets = Vec::with_capacity(aggregate_functions.len());
+            states_layout = Option::Some(get_layout_offsets(
+                aggregate_functions.as_slice(),
+                &mut states_offsets,
+            )?);
+        }
 
         let group_data_fields = group_expr
             .iter()
@@ -136,7 +153,15 @@ impl AggregatorParams {
             aggregate_functions_arguments_name.push(expr.to_aggregate_function_names()?);
         }
 
-        let (states_layout, states_offsets) = unsafe { get_layout_offsets(&aggregate_functions) };
+        let mut states_offsets: Vec<usize> = Vec::with_capacity(aggregate_functions.len());
+        let mut states_layout = None;
+        if !aggregate_functions.is_empty() {
+            states_offsets = Vec::with_capacity(aggregate_functions.len());
+            states_layout = Option::Some(get_layout_offsets(
+                aggregate_functions.as_slice(),
+                &mut states_offsets,
+            )?);
+        }
 
         let group_data_fields = plan
             .group_expr