--- a/query/src/pipelines/new/processors/transforms/aggregator/aggregator_partial.rs
+++ b/query/src/pipelines/new/processors/transforms/aggregator/aggregator_partial.rs
@@ -103,9 +103,10 @@ impl<const HAS_AGG: bool, Method: HashMethod + PolymorphicKeysHelper<Method> + S
 
             match inserted {
                 true => {
-                    let place = state.alloc_layout2(params);
-                    places.push(place);
-                    entity.set_state_value(place.addr());
+                    if let Some(place) = state.alloc_layout2(params) {
+                        places.push(place);
+                        entity.set_state_value(place.addr());
+                    }
                 }
                 false => {
                     let place: StateAddr = (*entity.get_state_value()).into();