--- a/query/src/pipelines/new/processors/transforms/aggregator/aggregator_partial.rs
+++ b/query/src/pipelines/new/processors/transforms/aggregator/aggregator_partial.rs
@@ -68,7 +68,7 @@ pub struct PartialAggregator<
 impl<const HAS_AGG: bool, Method: HashMethod + PolymorphicKeysHelper<Method> + Send>
     PartialAggregator<HAS_AGG, Method>
 {
-    pub fn create(method: Method, params: Arc<AggregatorParams>, ctx: Arc<QueryContext>) -> Self {
+    pub fn create(ctx: Arc<QueryContext>, method: Method, params: Arc<AggregatorParams>) -> Self {
         let state = method.aggregate_state();
         Self {
             is_generated: false,