I hereby agree to the terms of the CLA available at: https://databend.rs/dev/policies/cla/

## Summary

**1. Avoid undefine behavior**: 
In some cases (only group by or `select distinct`), in transform aggregation, `aggregate_functions` are empty, so in `get_layout_offsets`, the align of the generated `Layout` is 0, which will cause UB when allocating memory.

**2. Delete unnecessary unsafe**:
Replace `from_size_align_unchecked` with `from_size_align` and delete unsafe.


## Changelog

- Improvement

## Related Issues

Fixes #issue