@zhang2014 2022-05-13
```
In some cases (only group by or select distinct), in transform aggregation, aggregate_functions are empty, so in get_layout_offsets, the align of the generated Layout is 0, which will cause UB when allocating memory.
```
maybe this should not happen. When there is no aggregate function, we will not memory allocation.