@zhang2014 2022-05-13
Performance test is ok on macos:

before:

```
mysql> SELECT max(number) FROM numbers_mt(10000000000) GROUP BY number % 3;
+-------------+
| max(number) |
+-------------+
|  9999999999 |
|  9999999997 |
|  9999999998 |
+-------------+
3 rows in set (5.89 sec)
Read 10000000000 rows, 74.51 GiB in 5.886 sec., 1.7 billion rows/sec., 12.66 GiB/sec.

mysql> SELECT max(number), sum(number) FROM numbers_mt(10000000000) GROUP BY number % 3;
+-------------+----------------------+
| max(number) | sum(number)          |
+-------------+----------------------+
|  9999999999 | 16666666668333333333 |
|  9999999997 | 16666666661666666667 |
|  9999999998 | 16666666665000000000 |
+-------------+----------------------+
3 rows in set (7.49 sec)
Read 10000000000 rows, 74.51 GiB in 7.489 sec., 1.34 billion rows/sec., 9.95 GiB/sec.

mysql> SELECT max(number), sum(number) FROM numbers_mt(10000000000) GROUP BY number % 3, number % 4, number % 5 LIMIT 10;
+-------------+--------------------+
| max(number) | sum(number)        |
+-------------+--------------------+
|  9999999960 | 833333331666666660 |
|  9999999941 | 833333328500000006 |
|  9999999984 | 833333335666666668 |
|  9999999963 | 833333332166666661 |
|  9999999962 | 833333331999999994 |
|  9999999954 | 833333330666666664 |
|  9999999942 | 833333328666666672 |
|  9999999989 | 833333336500000003 |
|  9999999990 | 833333336666666670 |
|  9999999994 | 833333337333333338 |
+-------------+--------------------+
10 rows in set (9.88 sec)
Read 10000000000 rows, 74.51 GiB in 9.861 sec., 1.01 billion rows/sec., 7.56 GiB/sec.
```

after:
```
mysql> SELECT max(number) FROM numbers_mt(10000000000) GROUP BY number % 3;
+-------------+
| max(number) |
+-------------+
|  9999999999 |
|  9999999997 |
|  9999999998 |
+-------------+
3 rows in set (5.91 sec)
Read 10000000000 rows, 74.51 GiB in 5.899 sec., 1.7 billion rows/sec., 12.63 GiB/sec.


mysql> SELECT max(number), sum(number) FROM numbers_mt(10000000000) GROUP BY number % 3;
+-------------+----------------------+
| max(number) | sum(number)          |
+-------------+----------------------+
|  9999999999 | 16666666668333333333 |
|  9999999997 | 16666666661666666667 |
|  9999999998 | 16666666665000000000 |
+-------------+----------------------+
3 rows in set (7.53 sec)
Read 10000000000 rows, 74.51 GiB in 7.514 sec., 1.33 billion rows/sec., 9.92 GiB/sec.

mysql> SELECT max(number), sum(number) FROM numbers_mt(10000000000) GROUP BY number % 3, number % 4, number % 5 LIMIT 10;
+-------------+--------------------+
| max(number) | sum(number)        |
+-------------+--------------------+
|  9999999960 | 833333331666666660 |
|  9999999941 | 833333328500000006 |
|  9999999984 | 833333335666666668 |
|  9999999963 | 833333332166666661 |
|  9999999962 | 833333331999999994 |
|  9999999954 | 833333330666666664 |
|  9999999942 | 833333328666666672 |
|  9999999989 | 833333336500000003 |
|  9999999990 | 833333336666666670 |
|  9999999994 | 833333337333333338 |
+-------------+--------------------+
10 rows in set (9.97 sec)
Read 10000000000 rows, 74.51 GiB in 9.961 sec., 1 billion rows/sec., 7.48 GiB/sec.
```