@xudong963 2022-05-13
> maybe this should not happen. When there is no aggregate function, we will not memory allocation.

Seems `Bump` will still allocate layout, then call `layout.align()`, if align is 0, UB will happen.